Fix use of transaction parameters for threshold change transaction (#1903)
